[feat] #114 데이트 일정 상세 뷰 API 연동 (데이트 일정 상세 조회 API)

+++ app/src/main/java/org/sopt/dateroad/presentation/ui/timelinedetail/navigation/TimelineDetailNavigation.kt
@@ -30,7 +30,7 @@ fun NavGraphBuilder.timelineDetailGraph(
         TimelineDetailRoute(
             padding = padding,
             popBackStack = popBackStack,
-            dateId = dateId,
+            dateId = dateId.toLong(),
             dateType = dateType,
             sourceScreen = true
         )
